+++ drakkar/__init__.py
@@ -1,6 +1,6 @@
 """Drakkar — Kafka subprocess orchestration framework."""
 
-__version__ = '1.14.0'
+__version__ = '1.14.1'
 
 from drakkar.annotations import AnnotatorLike
 from drakkar.app import DrakkarApp